Create Stage zero
I made use of unpacking tuple method to solve the task problem. I assigned the solution to a variable called Introduction. Then I assigned the questions in the task to the variable Introduction also. Afterwards I printed it using F-string method. Which then printed the solution in an organised format.